EES-1557 Optimize glossary page bundle size
